LINE-BOT-DEMO/requirements.txt
Flask==2.2.2
click==8.1.3
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
googletrans==3.1.0a0
line-bot-sdk==2.1.0
requests
pandas
lxml
bs4
